refactor(gameState): 모델에서 서비스 dto로 패키지 이동
